plotting is only called from batch, so move it into batch to avoid cyclic graphs in Sigrid analysis between main files and batch files.
