+++ dfastmi/batch/ReporterDflowfm.py
@@ -39,7 +39,7 @@ from dfastmi.batch.SedimentationData import SedimentationData
 from dfastmi.batch.XykmData import XykmData
 from dfastmi.io.ApplicationSettingsHelper import ApplicationSettingsHelper
 from dfastmi.io.map_file import MapFile
-from dfastmi.plotting import plot_overview, savefig, zoom_xy_and_save
+from dfastmi.batch.plotting import plot_overview, savefig, zoom_xy_and_save
 
 
 class ReporterDflowfm:
